Inline Distro fetch formatting and drop the extra helper module.
Those helpers were only used by distro fetch, so they do not need their own file or tests.
